+++ allennlp/data/dataset_readers/conll2003.py
@@ -1,4 +1,4 @@
-from typing import Dict, List, Sequence, Iterable
+from typing import Dict, List, Sequence
 import itertools
 import logging
 
@@ -105,7 +105,7 @@ class Conll2003DatasetReader(DatasetReader):
         self._original_coding_scheme = "IOB1"
 
     @overrides
-    def _read(self, file_path: str) -> Iterable[Instance]:
+    def _read(self, file_path: str):
         # if `file_path` is a URL, redirect to the cache
         file_path = cached_path(file_path)
 
@@ -124,7 +124,12 @@ class Conll2003DatasetReader(DatasetReader):
                     # TextField requires `Token` objects
                     tokens = [Token(token) for token in tokens_]
 
-                    yield self.text_to_instance(tokens, pos_tags, chunk_tags, ner_tags)
+                    yield {
+                        "tokens": tokens,
+                        "pos_tags": pos_tags,
+                        "chunk_tags": chunk_tags,
+                        "ner_tags": ner_tags,
+                    }
 
     def text_to_instance(  # type: ignore
         self,