+++ CHANGELOG.md
@@ -23,6 +23,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - `Predictor.from_path` now automatically loads plugins (unless you specify `load_plugins=False`) so
   that you don't have to manually import a bunch of modules when instantiating predictors from
   an archive path.
+- Lazy dataset readers now work correctly with multi-process data loading.
 
 ### Added
 
@@ -48,6 +49,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - To be consistent with PyTorch `IterableDataset`, `AllennlpLazyDataset` no longer implements `__len__()`.
   Previously it would always return 1.
 - Removed old tutorials, in favor of [the new AllenNLP Guide](https://guide.allennlp.org)
+- When implementing the `DatasetReader._read` method, you should now return a `dict` of arguments
+  that will be passed to `DatasetReader.text_to_instance`, instead of an actual `Instance`.
+  For each `result` dictionary generated from `DatasetReader._read`, an actual `Instance`
+  will be created by calling `DatasetReader.text_to_instance(**result)`.
+  However, for backwards compatability you can still generate actual `Instance`s directly
+  from `_read`, but this probably be less efficient when using a multi-process data loader.
 
 ## [v1.0.0rc5](https://github.com/allenai/allennlp/releases/tag/v1.0.0rc5) - 2020-05-26